add nested roustes in planet_routes.py, routes_helper.py and modified tests.
