+++ app/routes/routes_helper.py
@@ -1,0 +1,16 @@
+from flask import abort, make_response
+
+# helper function 
+def validate_model(cls, model_id):
+    try:
+        model_id = int(model_id)
+    except:
+        abort(make_response({"message" : f" {cls.__name__} {model_id} invalid."}, 400))
+    
+    model = cls.query.get(model_id)
+    
+    if model:
+        return model
+        
+    abort(make_response({"message" : f" {cls.__name__} {model_id} not found."}, 404))
+
